Feature/omar/purchase inovice

+++ Pharamcy.Domain/Models/PartitionMedicineTracking.cs
@@ -10,12 +10,12 @@ namespace Pharamcy.Domain.Models
     {
         public string? BarCode { get; set; }
         public int Amount  => TabletsAvailableAmount / Tablets / Taps;
-        public decimal PurchasePrice { get; set; }
-        public decimal SalePrice { get; set; }
+        public double PurchasePrice { get; set; }
+        public double SalePrice { get; set; }
         public int Taps{ get; set; }
         public int Tablets{ get; set; }
         public int TabletsAvailableAmount { get; set; }
-        public decimal TabletSalePrice { get; set; }
+        public double TabletSalePrice { get; set; }
         public int PartitionMedicineId { get; set; }
         public virtual PartitionMedicine PartitionMedicine { get; set; }
     }
