+++ Pharamcy.Domain/Models/PruchaseInvoiceItem.cs
@@ -13,14 +13,14 @@ namespace Pharamcy.Domain.Models
         public int Amount { get; set; }
         public int Bonse { get; set; }
         public int Shelf { get; set; } = 0;
-        public decimal UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public int Taps { get; set; }
         public int Tablets { get; set; }
         public int Taxis { get; set; }
         public int AdditionalDiscount { get; set; }
         public int BaseDiscount { get; set; }
-        public decimal PurchasePrice { get; set; }
-        public decimal SalePrice { get; set; }
+        public double PurchasePrice { get; set; }
+        public double SalePrice { get; set; }
         public DateOnly ExpireDate { get; set; }
         public int PurchaseInvoiceId { get; set; }
         public virtual PurchaseInvoice PurchaseInvoice { get; set; }
